all users and fetching will add form for new users

+++ src/main/resources/static/mainquery.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS `users`;
 
 CREATE TABLE `users`
 (
-    `user_id`  BIGINT NOT NULL,
+    `user_id`  INT NOT NULL,
     `username` VARCHAR(50) NOT NULL,
     `password` VARCHAR(50),
     `active`   INT DEFAULT 1,
@@ -20,8 +20,8 @@ DROP TABLE IF EXISTS `roles`;
 
 CREATE TABLE `roles`
 (
-    `role_id` BIGINT NOT NULL auto_increment,
-    `user_id` BIGINT NOT NULL,
+    `role_id` INT NOT NULL auto_increment,
+    `user_id` INT NOT NULL,
     `role`    VARCHAR(50) NOT NULL DEFAULT "user",
     PRIMARY KEY(`role_id`),
     UNIQUE KEY `authorities_idx_1` (`user_id`, `role`),
@@ -34,7 +34,7 @@ DROP TABLE IF EXISTS `media`;
 
 CREATE TABLE `media`
 (
-    `media_id` BIGINT NOT NULL,
+    `media_id` INT NOT NULL,
     `title`    VARCHAR(256) DEFAULT "noname",
     PRIMARY KEY(`media_id`)
 )
@@ -44,8 +44,8 @@ DROP TABLE IF EXISTS `entry`;
 
 CREATE TABLE `entry`
 (
-    `user_id`  BIGINT NOT NULL,
-    `media_id` BIGINT NOT NULL,
+    `user_id`  INT NOT NULL,
+    `media_id` INT NOT NULL,
     `score`    INT,
     `status`   VARCHAR(20) NOT NULL,
     PRIMARY KEY `entries_idx_1` (`user_id`, `media_id`),
